Implemented bad data error handling
Put in a try/except to convert serial data to integer. This catches the
random bad serial data. Be nice to know the cause of it, but at least
it doesn't crash the program now. It seems this only happens after the
program has been exited once with a reupload of the arduino code. It
doesn't happen frequently during a run though.